refactor: 스페이스 삭제 domain 훅으로 분리
